move envoy BUILD top level
