+++ BUILD
@@ -14,6 +14,11 @@
 #
 ################################################################################
 #
+load("@rules_pkg//:pkg.bzl", "pkg_tar")
+load(
+    "@envoy//bazel:envoy_build_system.bzl",
+    "envoy_cc_binary",
+)
 
 exports_files(["LICENSE"])
 
@@ -22,5 +27,38 @@ config_setting(
     values = {
         "cpu": "darwin",
     },
+    visibility = ["//visibility:public"],
+)
+
+envoy_cc_binary(
+    name = "envoy",
+    repository = "@envoy",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//extensions/access_log_policy:access_log_policy_lib",
+        "//extensions/attributegen:attributegen_plugin",
+        "//extensions/metadata_exchange:metadata_exchange_lib",
+        "//extensions/stackdriver:stackdriver_plugin",
+        "//extensions/stats:stats_plugin",
+        "//source/extensions/filters/http/alpn:config_lib",
+        "//source/extensions/filters/http/authn:filter_lib",
+        "//source/extensions/filters/http/istio_stats",
+        "//source/extensions/filters/listener/set_internal_dst_address:filter_lib",
+        "//source/extensions/filters/network/forward_downstream_sni:config_lib",
+        "//source/extensions/filters/network/istio_authn:config_lib",
+        "//source/extensions/filters/network/metadata_exchange:config_lib",
+        "//source/extensions/filters/network/sni_verifier:config_lib",
+        "//source/extensions/filters/network/tcp_cluster_rewrite:config_lib",
+        "@envoy//source/exe:envoy_main_entry_lib",
+    ],
+)
+
+pkg_tar(
+    name = "envoy_tar",
+    srcs = [":envoy"],
+    extension = "tar.gz",
+    mode = "0755",
+    package_dir = "/usr/local/bin/",
+    tags = ["manual"],
     visibility = ["//visibility:public"],
 )
